+++ sdk/freeport/systemlimit_windows.go
@@ -2,6 +2,6 @@
 
 package freeport
 
-func systemLimit() (uint64, error) {
+func systemLimit() (int, error) {
 	return 0, nil
 }
